server/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.7.0
eventlet==0.37.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.4.1
fsspec==2024.9.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.10
langchain-openai==0.2.2
langsmith==0.1.134
MarkupSafe==3.0.1
numpy==2.1.2
openai==1.51.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.0
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
